restructured database
fixed count button
